requirements.txt
molo.core
django-contrib-comments==1.7.1
django-mptt
django-import-export
wagtailmodeladmin
